requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
bitsandbytes>=0.38.1
datasets>=2.0.0
accelerate>=0.18.0
opencv-python>=4.7.0
Pillow>=9.2.0
fastapi>=0.88.0
uvicorn>=0.22.0
requests>=2.28.0